Move bot functions to lib
