+++ src/main/kotlin/com/ooooonly/luaMirai/lua/MiraiGlobals.kt
@@ -89,76 +89,12 @@ open class MiraiGlobals() : Globals() {
         LuaC.install(this)
 
         initEventTable()
-        initMsgConstructor()
-        initBotFactory()
     }
 
     private fun initEventTable() = set("Event", LuaTable())
-    private fun initMsgConstructor() {
-        set("Msg", object : VarArgFunction() {
-            override fun onInvoke(args: Varargs): Varargs {
-                val arg1 = args.arg1()
-                if (arg1 is LuaTable) return MiraiMsg(arg1)
-                else if (arg1 is LuaString) return MiraiMsg(arg1.toString())
-                return MiraiMsg()
-            }
-        })
-        setFunction1Arg("At") {
-            it.takeIfType<MiraiGroupMember>()?.let {
-                MiraiMsg(At(it.member))
-            } ?: MiraiMsg()
-        }
 
-        setFunction1Arg("Quote") {
-            when (it) {
-                is MiraiMsg -> it.chain[MessageSource]?.let { MiraiMsg(QuoteReply(it)) } ?: MiraiMsg()
-                is MiraiSource -> it.source?.let { MiraiMsg(QuoteReply(it)) }
-                else -> MiraiMsg()
-            }
-        }
 
-        setFunction2Arg("Image") { arg1, arg2 ->
-            MiraiMsg.getImage(
-                arg1,
-                arg2.takeIf { it != LuaValue.NIL }
-            )?.let { MiraiMsg(it) } ?: MiraiMsg()
-        }
 
-        setFunction2Arg("UploadImage") { arg1, arg2 ->
-            MiraiMsg.uploadImage(
-                arg1,
-                arg2.takeIf { it != LuaValue.NIL }
-            )?.let { MiraiMsg(it) } ?: MiraiMsg()
-        }
-
-        setFunction2Arg("ImageFile") { arg1, arg2 ->
-            MiraiMsg.uploadImage(
-                File(arg1.checkjstring()).toURI().toURL().toString().toLuaValue(),
-                arg2.takeIf { it != LuaValue.NIL }
-            )?.let { MiraiMsg(it) } ?: MiraiMsg()
-        }
-
-        setFunction0Arg("AtAll") {
-            MiraiMsg(AtAll)
-        }
-
-        setFunction1Arg("Face") {
-            MiraiMsg(Face(it.checkint()))
-        }
-    }
-
-    private fun initBotFactory() {
-        setFunction2Arg("Bot") { user, password ->
-            return@setFunction2Arg MiraiBot(user.checklong(), password.checkjstring(), 0)
-        }
-        setFunction("Bot") {
-            val user = it.checklong(1)
-            val pwd = it.checkjstring(2)
-            val config =
-                if (it.narg() >= 3) BotConfiguration.Default.apply { fileBasedDeviceInfo(it.checkjstring(3)) } else BotConfiguration.Default
-            return@setFunction MiraiBot(user, pwd, 0, config)
-        }
-    }
 
     fun onLoad(bot: Bot) = onLoadFun?.let {
         mBot = MiraiBot(bot, idCounter++)
